Add devtools
