+++ app/main.js
@@ -21,18 +21,12 @@ runLoop.setStore(store);
 runLoop.start();
 
 // Set up router
-import {Router} from 'react-router';
 
 import routes from './config/routes';
 import history from './config/history';
 
-// Render root
-import {Provider} from 'react-redux';
+import Root from './views/Root';
 
 ReactDOM.render((
-  <Provider store={store}>
-    <Router history={history}>
-      {routes}
-    </Router>
-  </Provider>
+  <Root store={store} routes={routes} history={history} />
 ), document.getElementById('container'));
