+++ app/views/App.js
@@ -2,11 +2,7 @@ import React from 'react';
 
 const App = React.createClass({
   render() {
-    return (
-      <div className="root">
-        {this.props.children}
-      </div>
-    );
+    return this.props.children;
   }
 });
