+++ package.json
@@ -47,6 +47,10 @@
     "minimatch": "^3.0.0",
     "mocha": "^2.3.4",
     "react-addons-test-utils": "^0.14.6",
+    "redux-devtools": "^3.0.1",
+    "redux-devtools-dock-monitor": "^1.0.1",
+    "redux-devtools-filter-actions": "^1.0.0",
+    "redux-devtools-log-monitor": "^1.0.2",
     "style-loader": "^0.12.2",
     "webpack": "^1.12.9",
     "webpack-dev-server": "^1.14.0",
